requirements.txt
Flask
beautifulsoup4
llama-cpp-python